+++ src/game/player/libElements.ts
@@ -1,6 +1,6 @@
 import Isolated from "@player/lib/isolated";
 import Say from "@player/elements/say/Say";
-import {Top, Center, Bottom} from "@player/lib/PlayerFrames";
+import {Top, Center, Bottom, Full} from "@player/lib/PlayerFrames";
 import {VBox, HBox} from "@player/lib/UtilComponents";
 
 export {
@@ -11,4 +11,5 @@ export {
     Bottom,
     VBox,
     HBox,
+    Full,
 };