+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `image.setTags`
 - `image.setPosition`
+- Utility component: `Full`
 
 ## [0.2.2] - 2024/12/02